% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{calcBHI}
\alias{calcBHI}
\title{Calculate "biological homogeneity index" from GO terms and uniprot accessn names}
\usage{
calcBHI(Accs, gos)
}
\description{
Calculate "biological homogeneity index" from GO terms and uniprot accessn names
}
