% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{mfuzz.plot}
\alias{mfuzz.plot}
\title{Plotting clustering results into multiple figure panels, adopted from the MFuzz package}
\usage{
mfuzz.plot(
  dat,
  cl,
  mfrow = c(1, 1),
  colo,
  min.mem = 0,
  time.labels,
  filename = NA,
  xlab = "Time",
  ylab = "Expression changes"
)
}
\arguments{
\item{filename}{for writing into pdf. Will write on screen when using NA}
}
\description{
Plotting clustering results into multiple figure panels, adopted from the MFuzz package
}
