% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{runFuncEnrich}
\alias{runFuncEnrich}
\title{"RGD_ID", "SGD_ID", "TAIR_ID", "UCSC_GENE_ID", "UNIGENE",}
\usage{
runFuncEnrich(cl, protnames = NULL, idtypes, infosource)
}
\description{
"RGD_ID", "SGD_ID", "TAIR_ID", "UCSC_GENE_ID", "UNIGENE",
}
