% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{SwitchOrder}
\alias{SwitchOrder}
\title{arrange cluster member numbers from largest to smallest}
\usage{
SwitchOrder(Bestcl, NClust)
}
\arguments{
\item{Bestcl}{fclust object}

\item{NClust}{Number of clusters}
}
\description{
arrange cluster member numbers from largest to smallest
}
