% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{vsclust_algorithm}
\alias{vsclust_algorithm}
\title{Run the vsclust clustering algorithm}
\usage{
vsclust_algorithm(
  x,
  centers,
  iter.max = 100,
  verbose = FALSE,
  dist = "euclidean",
  m = 2,
  rate.par = NULL,
  weights = 1,
  control = list()
)
}
\arguments{
\item{x}{a numeric data matrix}

\item{centers}{Either numeric for number of clusters or numeric matrix with center coordinates}

\item{iter.max}{Numeric for maximum number of iterations}

\item{verbose}{Verbose information}

\item{m}{Fuzzifier value: numeric or vector of length equal to number of rows of x}

\item{rate.par}{(experimental) numeric value for punishing missing values}

\item{weights}{numeric or vector of length equal to number of rows of x}

\item{control}{list with arguments to vsclust algorithms (now only cutoff for relative tolerance: reltol)}
}
\value{
Xie Beni index
}
\description{
Run the vsclust clustering algorithm
}
\examples{
TODO
}
