% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{ClustComp}
\alias{ClustComp}
\title{Running VSClust on given data set (with data pre-processing)}
\usage{
ClustComp(tData, NSs = 10, NClust = NClust, Sds = Sds, cores = 1)
}
\description{
Running VSClust on given data set (with data pre-processing)
}
