% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{cvalidate.xiebeni}
\alias{cvalidate.xiebeni}
\title{Calculate the Xie Beni index for validation of the cluster number}
\usage{
cvalidate.xiebeni(clres, m)
}
\arguments{
\item{clres}{Output from clustering. Either fclust object or similar list}

\item{m}{Fuzzifier value}
}
\value{
Xie Beni index
}
\description{
Calculate the Xie Beni index for validation of the cluster number
}
\examples{
TODO
}
